refactor(web): split schedules page helpers [OPE-364]
